+++ api/src/main/java/com/canhlabs/funnyapp/cache/MFASessionStore.java
@@ -1,28 +1,11 @@
 package com.canhlabs.funnyapp.cache;
 
-import org.springframework.stereotype.Service;
-
 import java.util.Optional;
 
-@Service
-public class MFASessionStore {
+public interface MFASessionStore {
+    void storeSession(String sessionToken, String userId);
 
-    // key: sessionToken, value: userId
-    private final AppCache<String, String> sessionCache;
+    Optional<String> getUserId(String sessionToken);
 
-    public MFASessionStore(AppCacheFactory factory) {
-        this.sessionCache = factory.createCache(5, 10000); // TTL: 5m
-    }
-
-    public void storeSession(String sessionToken, String userId) {
-        sessionCache.put(sessionToken, userId);
-    }
-
-    public Optional<String> getUserId(String sessionToken) {
-        return sessionCache.get(sessionToken);
-    }
-
-    public void remove(String sessionToken) {
-        sessionCache.invalidate(sessionToken);
-    }
+    void remove(String sessionToken);
 }
