+++ api/src/test/java/com/canhlabs/funnyapp/service/impl/ConverterTest.java
@@ -1,5 +1,5 @@
 // File: src/test/java/com/canhlabs/funnyapp/share/ConverterTest.java
-package com.canhlabs.funnyapp.share;
+package com.canhlabs.funnyapp.service.impl;
 
 import com.canhlabs.funnyapp.domain.ShareLink;
 import com.canhlabs.funnyapp.domain.User;
