+++ api/src/test/java/com/canhlabs/funnyapp/service/impl/ShareServiceImplTest.java
@@ -7,7 +7,6 @@ import com.canhlabs.funnyapp.repo.ShareLinkRepo;
 import com.canhlabs.funnyapp.repo.UserRepo;
 import com.canhlabs.funnyapp.config.AppProperties;
 import com.canhlabs.funnyapp.share.AppUtils;
-import com.canhlabs.funnyapp.share.Converter;
 import com.canhlabs.funnyapp.dto.ShareRequestDto;
 import com.canhlabs.funnyapp.dto.UserDetailDto;
 import com.canhlabs.funnyapp.dto.VideoDto;
